Expect WatchPoint::cancel() to remove the watcher already
On Alpine Linux in a Docker container we've seen that it is possible that calling inotify_rm_watch() immediately produces the IN_IGNORED event, which in turn can result in the watch point to be deleted before returning from the cancel() method, resulting in a crash.

This change ensures that we ignore already removed watch points instead.